+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.2.21](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.2.21)
+### Changed
+- Modify LEARCredentialMachine issuance policy validations to enforce that non-admins can only include `ProductOffering`.
+- Modify LEARCredentialEmployee and LEARCredentialMachine issuance to enforce that there is always at least one power included for all users.
+
 ## [v2.2.20](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.2.20)
 ### Changed
 - Send label credential to response URI in parallel to sending the credential offer email.